+++ app/api/v1/routes_reviews.py
@@ -1,30 +1,55 @@
 from flask_restx import Namespace, Resource, fields
 from flask import current_app
+from flask_jwt_extended import jwt_required, get_jwt_identity # type: ignore
 
 
 api = Namespace('reviews', description='Review operations')
 
 # Define the review model for input validation and documentation
 review_model = api.model('Review', {
-    'comment': fields.String(required=True, description='Text of the review'),
+    'text': fields.String(required=True, description='Text of the review'),
     'rating': fields.Integer(required=True, description='Rating of the place (1-5)'),
     'user_id': fields.String(required=True, description='ID of the user'),
     'place_id': fields.String(required=True, description='ID of the place')
 })
 
+auth_header = {'Authorization': {
+        'description': 'Bearer <JWT Token>',
+        'in': 'header',
+        'type': 'string',
+        'required': True
+    }
+}
 
 @api.route('/')
 class ReviewList(Resource):
     @api.expect(review_model)
     @api.response(201, 'Review successfully created')
     @api.response(400, 'Invalid input data')
+    @api.doc('update_user', params=auth_header)
+    @jwt_required()
     def post(self):
         """Register a new review"""
-        facade = current_app.facade # type: ignore
+        facade = current_app.extensions['FACADE']
+        current_user = get_jwt_identity()
         review_data = api.payload
 
+        place = facade.get_place(review_data['place_id'])
+
+        #verify login
+        if review_data["user_id"] != current_user["id"]:
+            return {'error': 'Unauthorized action'}, 403
+        #verify if place is owned by reviewer
+        if place.owner_id == current_user["id"]:
+            return {'error': 'Unauthorized action: You cannot review your own place.'}, 403
+        #verify if place already reviewed:
+        reviews = facade.get_all_reviews()
+        for review in reviews:
+            if current_user["id"] == review.user_id and place.id == review.place_id:
+                return {'error': 'Unauthorized action: You already reviewed this place.'}, 403
+
         # Validate required fields
-        if not all(key in review_data for key in ('user_id', 'place_id', 'rating', 'comment')):
+        if not all(key in review_data for key in ('user_id', 'place_id', 'rating', 'text')):
             return {'error': 'Missing required fields'}, 400
 
         # Validate user and place exist
@@ -43,7 +68,7 @@ class ReviewList(Resource):
 
         return {
             'id': new_review.id,
-            'comment': new_review.comment,
+            'text': new_review.text,
             'rating': new_review.rating,
             'user_id': new_review.user_id,
             'place_id': new_review.place_id
@@ -52,13 +77,13 @@ class ReviewList(Resource):
     @api.response(200, 'List of reviews retrieved successfully')
     def get(self):
         """Retrieve a list of all reviews"""
-        facade = current_app.facade # type: ignore
+        facade = current_app.extensions['FACADE']
         reviews = facade.get_all_reviews()
         review_list = []
         for review in reviews:
             review_list.append({
                 'id': review.id,
-                'comment': review.comment,
+                'text': review.text,
                 'rating': review.rating,
                 'user_id': review.user_id,
                 'place_id': review.place_id
@@ -71,13 +96,13 @@ class ReviewResource(Resource):
     @api.response(404, 'Review not found')
     def get(self, review_id):
         """Get review details by ID"""
-        facade = current_app.facade # type: ignore
+        facade = current_app.extensions['FACADE']
         review = facade.get_review(review_id)
         if not review:
             return {'error': 'Review not found'}, 404
         return {
             'id': review.id,
-            'comment': review.comment,
+            'text': review.text,
             'rating': review.rating,
             'user_id': review.user_id,
             'place_id': review.place_id
@@ -87,14 +112,27 @@ class ReviewResource(Resource):
     @api.response(200, 'Review updated successfully')
     @api.response(404, 'Review not found')
     @api.response(400, 'Invalid input data')
+    @api.doc('update_user', params=auth_header)
+    @jwt_required()
     def put(self, review_id):
         """Update a review's information"""
-        facade = current_app.facade # type: ignore
+        current_user = get_jwt_identity()
+        is_admin = current_user.get('is_admin', False)
+        facade = current_app.extensions['FACADE']
+
         review = facade.get_review(review_id)
         if not review:
             return {'error': 'Review not found'}, 404
+        
+        #verify login
+        if not is_admin and review.user_id != current_user["id"]:
+            return {'error': 'Unauthorized action, you can only update your own reviews'}, 403
 
         review_data = api.payload
+
+        #verify login
+        if not is_admin and review_data["user_id"] != current_user["id"]:
+            return {'error': 'Unauthorized action, a review must contain your id'}, 403
 
         try:
             facade.update_review(review_id, review_data)
@@ -104,7 +142,7 @@ class ReviewResource(Resource):
         updated_review = facade.get_review(review_id)
         return {
             'id': updated_review.id,
-            'comment': updated_review.comment,
+            'text': updated_review.text,
             'rating': updated_review.rating,
             'user_id': updated_review.user_id,
             'place_id': updated_review.place_id
@@ -112,12 +150,21 @@ class ReviewResource(Resource):
 
     @api.response(200, 'Review deleted successfully')
     @api.response(404, 'Review not found')
+    @api.doc('update_user', params=auth_header)
+    @jwt_required()
     def delete(self, review_id):
         """Delete a review"""
-        facade = current_app.facade # type: ignore
+        current_user = get_jwt_identity()
+        is_admin = current_user.get('is_admin', False)
+        facade = current_app.extensions['FACADE']
+
         review = facade.get_review(review_id)
         if not review:
             return {'error': 'Review not found'}, 404
+        
+        #verify login
+        if not is_admin and review.user_id != current_user["id"]:
+            return {'error': 'Unauthorized action, you can only delete your own reviews'}, 403
 
         facade.delete_review(review_id)
         return {'message': f'Review {review_id} deleted successfully'}, 200
@@ -128,7 +175,7 @@ class PlaceReviewList(Resource):
     @api.response(404, 'Place not found')
     def get(self, place_id):
         """Get all reviews for a specific place"""
-        facade = current_app.facade # type: ignore
+        facade = current_app.extensions['FACADE']
         place = facade.get_place(place_id)
         if not place:
             return {'error': 'Place not found'}, 404
@@ -138,7 +185,7 @@ class PlaceReviewList(Resource):
         for review in reviews:
             reviews_list.append({
                 'id': review.id,
-                'comment': review.comment,
+                'text': review.text,
                 'rating': review.rating,
                 'user_id': review.user_id,
                 'place_id': review.place_id
